requirements.txt
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
dnspython==2.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
mysql-connector-python==9.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
playwright==1.47.0
pluggy==1.5.0
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
pymongo==4.9.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.8.1
selenium==4.25.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
